requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.1.0
atari-py==0.1.7
-e git+git@github.com:DrKwint/baselines.git@bcaa0e6cabd4283ef6432282e7347b234aa6b2c5#egg=baselines
box2d-py==2.3.8
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29.2
decorator==4.3.0
dill==0.2.8.2
future==0.17.1
gast==0.2.0
glfw==1.7.0
grpcio==1.17.1
gym==0.10.9
h5py==2.9.0
idna==2.8
imageio==2.4.1
isort==4.3.4
joblib==0.13.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lockfile==0.12.2
Markdown==3.0.1
matplotlib==3.0.2
mccabe==0.6.1
mock==2.0.0
mujoco-py==1.50.1.68
networkx==2.2
numpy==1.15.4
opencv-python==3.4.5.20
pandas==0.23.4
pbr==5.1.1
Pillow==5.4.0
progressbar2==3.39.2
protobuf==3.6.1
pycparser==2.19
pyglet==1.3.2
pygraphviz==1.5
pylint==2.2.2
PyOpenGL==3.1.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
requests==2.21.0
scipy==1.2.0
six==1.12.0
tensorboard==1.12.1
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.28.1
typed-ast==1.1.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
yapf==0.25.0
